auth-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.3
prometheus-flask-exporter==0.23.1
